requirements.txt
yt-dlp
customtkinter
Pillow
packaging
pyinstaller
